Modularize client game logic
