+++ package.json
@@ -4,7 +4,8 @@
   "description": "",
   "main": "server.js",
   "scripts": {
-    "start": "node server.js"
+    "start": "node server.js",
+    "test": "jest"
   },
   "repository": {
     "type": "git",
@@ -32,5 +33,9 @@
   "engines": {
     "node": ">=16.0.0 <17.0.0",
     "npm": ">=8.0.0 <9.0.0"
+  },
+  "devDependencies": {
+    "jest": "^29.7.0",
+    "jsdom": "^21.1.0"
   }
 }
