+++ views/index.html
@@ -29,7 +29,12 @@
 	<script src="https://cdn.socket.io/3.1.1/socket.io.min.js"></script>
 	<script type="text/javascript" src="/javascripts/jquery.min.js"></script>
 	<script type="text/javascript" src="/javascripts/phaser.min.js"></script>
-	<script type="text/javascript" src="/javascripts/game1.js"></script>
+        <script type="text/javascript" src="/javascripts/states/landing.js"></script>
+        <script type="text/javascript" src="/javascripts/states/gameplay.js"></script>
+        <script type="text/javascript" src="/javascripts/states/gameover.js"></script>
+        <script type="text/javascript" src="/javascripts/states/leaderboard.js"></script>
+        <script type="text/javascript" src="/javascripts/states/instructions.js"></script>
+        <script type="text/javascript" src="/javascripts/game.js"></script>
 </div>
 
 <label></label>
